can't use a symlink in the gem
the link is broken when the gem is installed, since the link target isn't
packaged. going to try reversing this and see how it goes. (i think symlinks
will work on github.com, but we'll see.)

if i can't make a symlink work, i'll have to have a build step which copies
rubocop.yml into the gem, but i'd prefer not to.