our gem now depends on the lastest rubocop release
any project which adopts this gem will then instal this version as a dependency.

it's possible this version specification will be overly strict. but rubocop
changes so much my goal is to pin to 1 version that we know our styles are
compatible with. loosening may be ok-ish, but only if this strictness proves to
be a problem and no other solution can be found. (I want to stop seeing
"outdated style configuration" warnings when I run rubocop.)